pyproject.toml
[tool.poetry]
name = "test-data-innvestigate"
version = "0.1.0"
description = "Test data for LRP implementations"
authors = ["Adrian Hill <hill@campus.tu-berlin.de>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/adrhill/test-data-innvestigate"
repository = "https://github.com/adrhill/test-data-innvestigate"
keywords = ["LRP", "XAI", "testing"]

[tool.poetry.dependencies]
# Versions specified for iNNvestigate 1.0.9 test
python = "3.6.*"
innvestigate = "1.0.9"
Keras = "2.2.4"
tensorflow = "1.12.*"

# Ensure compatibility of numpy with TF 1.12
numpy = "1.16.*"

# Other packages
h5py = "^3.1.0"
jupyter = "^1.0.0"
matplotlib = "^3.3.4"
Pillow = "^8.1.1"
tqdm = "^4.58.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black  = { version = "*", allow-prereleases = true }
flake8 = "^3.8.4"
flake8-black = "^0.2.1"
flake8-bugbear = "^21.3.1"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.5.0"
pep8-naming = "^0.11.1"
pytest = "^6.2.2"
rope = "^0.18.0"
isort = "^5.8.0"
flake8-isort = "^4.0.0"
pylint = "^2.8.2"

[tool.isort]
profile = "black"
multi_line_output = 3 
known_first_party = "innvestigate"
force_single_line = "True"
group_by_package = "True"
force_alphabetical_sort_within_sections = "True"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::FutureWarning:tensorflow",
    "ignore::DeprecationWarning:tensorflow",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test-data-innvestigate = "test_data_innvestigate.runall:main"
